transaction controllers outline

+++ db/poke_mon_ey.sql
@@ -16,7 +16,7 @@ CREATE TABLE merchants (
 
 CREATE TABLE transactions (
     id SERIAL PRIMARY KEY,
-    amount FLOAT,
+    amount INT,
     tag_id INT REFERENCES tags(id) ON DELETE CASCADE,
     merchant_id INT REFERENCES merchants(id) ON DELETE CASCADE
 );
